requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
coverage==7.6.12
dnspython==2.7.0
ecdsa==0.19.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.25.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.14.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.24.0
